LayoutTests/http/tests/security/contentSecurityPolicy/manifest-src-allowed.html

<meta http-equiv="Content-Security-Policy" content="manifest-src *">
<link rel="manifest" href="manifest.test/manifest.json">
<script>
    if (window.testRunner) {
        testRunner.dumpAsText();
        testRunner.waitUntilDone();
        (async () => {
            const success = await testRunner.getApplicationManifest();
            alert("manifest fetch success: " + success);
            testRunner.notifyDone();
        })();
    }
</script>
