LayoutTests/http/tests/security/contentSecurityPolicy/manifest-src-allowed-expected.txt

ALERT: manifest fetch success: true
